pyproject.toml
[tool.black]
line_length = 79

[tool.coverage.run]
branch = true
source = ["rororo"]

[tool.coverage.paths]
source = ["rororo/"]

[tool.coverage.report]
fail_under = 95
skip_covered = true
show_missing = true

[tool.poetry]
name = "rororo"
version = "2.0.2a0"
description = "aiohttp.web OpenAPI 3 schema first server applications."
authors = ["Igor Davydenko <iam@igordavydenko.com>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://igordavydenko.com/projects/#rororo"
repository = "https://github.com/playpauseandstop/rororo"
documentation = "https://rororo.readthedocs.io/"
packages = [
  {include = "rororo", from="src"},
]
keywords = ["aiohttp", "aiohttp.web", "oas", "openapi", "openapi3", "helpers", "utilities"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "Framework :: AsyncIO",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet :: WWW/HTTP :: WSGI",
  "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
  "Topic :: Utilities",
  "Typing :: Typed"
]

[tool.poetry.dependencies]
python = "^3.6"
aiohttp = "^3.5"
aiohttp-middlewares = "^1.1.0"
attrs = ">=19.1,<21"
email-validator = "^1.0.5"
environ-config = "^20.1.0"
isodate = "^0.6.0"
openapi-core = "^0.13.3"
pyrsistent = "^0.16.0"
PyYAML = "^5.1"
typing-extensions = { version = "^3.7.4", python = "<3.8" }

[tool.poetry.dev-dependencies]
aioredis = "^1.3.1"
async_generator = {version = "^1.10", python = "~3.6"}
coverage = {extras = ["toml"], version = "^5.0.3"}
pytest = "^6.0.1"
pytest-aiohttp = "^0.3.0"
pytest-cov = "^2.7"
pytest-env = "^0.6.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/playpauseandstop/rororo/issues"

[tool.pytest.ini_options]
minversion = "6.0.1"
testpaths = [
  "examples/hobotnica/tests.py",
  "examples/petstore/tests.py",
  "examples/simulations/tests.py",
  "examples/todobackend/tests.py",
  "tests/",
]
addopts = "--cov --no-cov-on-fail"
env = """
LEVEL=test
"""
log_level = "info"

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py36,py37,py38,py38-prev
skipsdist = True

[gh-actions]
python =
  3.6: py36
  3.7: py37
  3.8: py38

[testenv]
passenv =
  PYTHONPATH
  PYTEST_ADDOPTS
  REDIS_URL
setenv =
  PYTHONPATH=.
  USER=playpauseandstop
skip_install = True
whitelist_externals = poetry
commands_pre =
  poetry install
commands =
  poetry run python -m pytest

[testenv:py38-prev]
commands_pre =
  poetry install
  poetry run python -m pip install attrs==19.3.0 openapi-core==0.13.3
"""
